Fix synchronization in CUDA tests.
For "generic" channel tests, we wrap an std::vector into a
tensorpipe::CudaBuffer by allocating a CUDA buffer of the same size
and calling cudaMemcpy(). We also create a non-blocking stream for
this buffer's use.
The issue here is that the stream being non-blocking, it can advance
before the copy actually starts, which manifested in the following
diff. Replacing the cudaMemcpy with a cudaMemcpyAsync enqueued on the
stream fixes the issue, and corresponds to the intended use of CUDA
buffers in TensorPipe. I am not certain why it hasn't manifested in
tests for other CUDA channels though.

ghstack-source-id: 76ae0dbf19b6dcb2fe52ce7a6bd4157c3139f7e7
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/302